wip: package playground as library
