+++ playground/src/main.tsx
@@ -1,12 +1,9 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import App from "./App.tsx";
-import "graphiql/graphiql.css";
-import "./index.css";
-import "@graphiql/plugin-explorer/dist/style.css";
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import App from './App.tsx';
 
-ReactDOM.createRoot(document.getElementById("root")!).render(
+ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>
+  </React.StrictMode>,
 );
