+++ playground/tsconfig.json
@@ -20,10 +20,12 @@
     "noUnusedParameters": false,
     "noFallthroughCasesInSwitch": true,
 
+    "typeRoots": ["./dist/index.d.ts", "node_modules/@types"],
+
     "paths": {
       "@/*": ["./src/*"]
     }
   },
-  "include": ["src"],
+  "include": ["src", "./index.ts"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
